refactor: streamline keysigning navigation and remove unused components
- Updated routing paths to use corePaths for keysigning, enhancing consistency across the application.
- Removed the StartFastKeysignPage component and integrated its functionality into the new StartKeysignProviders for better modularity.
- Replaced the use of the deprecated keysignMessagePayload with corePathState for improved state management.
- Cleaned up imports and eliminated unused components to enhance code organization and maintainability.
