+++ clients/desktop/src/vault/deposit/DepositConfirmButton/index.tsx
@@ -5,6 +5,7 @@ import { coinKeyFromString } from '@core/chain/coin/Coin'
 import { toCommCoin } from '@core/mpc/types/utils/commCoin'
 import { KeysignPayloadSchema } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
 import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
+import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
 import {
   useCurrentVault,
   useCurrentVaultSecurityType,
@@ -17,7 +18,6 @@ import { shouldBePresent } from '@lib/utils/assert/shouldBePresent'
 import { useTranslation } from 'react-i18next'
 
 import { toHexPublicKey } from '../../../chain/utils/toHexPublicKey'
-import { useAppNavigate } from '../../../navigation/hooks/useAppNavigate'
 import { useAppPathParams } from '../../../navigation/hooks/useAppPathParams'
 import { useVaultPublicKeyQuery } from '../../publicKey/queries/useVaultPublicKeyQuery'
 import { useCurrentVaultCoin } from '../../state/currentVaultCoins'
@@ -25,8 +25,6 @@ import { ChainAction } from '../ChainAction'
 import { useCurrentDepositCoin } from '../hooks/useCurrentDepositCoin'
 import { useDepositChainSpecificQuery } from '../queries/useDepositChainSpecificQuery'
 import { transactionConfig } from './config'
-
-type DepositType = 'fast' | 'paired'
 
 type DepositConfirmButtonProps = {
   depositFormData: Record<string, unknown>
@@ -43,7 +41,7 @@ export const DepositConfirmButton = ({
   const { t } = useTranslation()
   const [coinKey] = useCurrentDepositCoin()
   const coin = useCurrentVaultCoin(coinKey)
-  const navigate = useAppNavigate()
+  const navigate = useCoreNavigate()
   const chainSpecificQuery = useDepositChainSpecificQuery()
   const vault = useCurrentVault()
   const config = transactionConfig[action] || {}
@@ -64,7 +62,7 @@ export const DepositConfirmButton = ({
 
   const walletCore = useAssertWalletCore()
 
-  const startKeysign = (type: DepositType) => {
+  const startKeysign = () => {
     // TODO: handle affiliate fee and percentage
     const publicKey = shouldBePresent(publicKeyQuery.data)
     const keysignPayload = create(KeysignPayloadSchema, {
@@ -94,7 +92,7 @@ export const DepositConfirmButton = ({
       ).toString()
     }
 
-    navigate(type === 'fast' ? 'fastKeysign' : 'keysign', {
+    navigate('keysign', {
       state: {
         keysignPayload: { keysign: keysignPayload },
       },
@@ -120,13 +118,13 @@ export const DepositConfirmButton = ({
   if (securityType === 'fast') {
     return (
       <VStack gap={20}>
-        <Button onClick={() => startKeysign('fast')}>{t('fast_sign')}</Button>
-        <Button kind="outlined" onClick={() => startKeysign('paired')}>
+        <Button onClick={startKeysign}>{t('fast_sign')}</Button>
+        <Button kind="outlined" onClick={startKeysign}>
           {t('paired_sign')}
         </Button>
       </VStack>
     )
   }
 
-  return <Button onClick={() => startKeysign('paired')}>{t('continue')}</Button>
+  return <Button onClick={startKeysign}>{t('continue')}</Button>
 }
