--- clients/desktop/src/vault/keysign/shared/state/keysignMessagePayload.tsx
@@ -1,7 +1,0 @@
-import { KeysignMessagePayload } from '@core/mpc/keysign/keysignPayload/KeysignMessagePayload'
-import { getValueProviderSetup } from '@lib/ui/state/getValueProviderSetup'
-
-export const {
-  useValue: useKeysignMessagePayload,
-  provider: KeysignMessagePayloadProvider,
-} = getValueProviderSetup<KeysignMessagePayload>('KeysignMessagePayload')
